requirements.txt
requests==2.31.0
python-telegram-bot==13.7
python-dotenv==1.0.0
tabulate==0.9.0
cronitor
